// frontend/src/api/model/labelClass.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Severity } from './severity';


/**
 * Label Object which includes start end and id  name: name of the Label color: color of the Label range: range of the Label
 */
export interface LabelClass { 
    name?: string;
    color?: string;
    severity?: Severity;
    id: string;
    priority: number;
}