// frontend/src/api/model/predictionRequest.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PredictionAlgorithm } from './predictionAlgorithm';


export interface PredictionRequest { 
    algorithm?: PredictionAlgorithm;
    params?: object;
}