// frontend/src/api/model/project.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IntegrationStatus } from './integrationStatus';


/**
 * Project - Model to pass to frontend     
 */
export interface Project { 
    name?: string;
    samples?: number;
    dimensions?: number;
    hasTimestamps?: boolean;
    status?: IntegrationStatus;
    id: string;
}