// frontend/src/api/model/severity.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 */
export type Severity = 'okay' | 'warning' | 'error';

export const Severity = {
    Okay: 'okay' as Severity,
    Warning: 'warning' as Severity,
    Error: 'error' as Severity
};